Category Servislerinin Düzenlenmesi

+++ Core/CafeAPI.Application/Services/Concrete/CategoryServices.cs
@@ -9,30 +9,68 @@ using CafeAPI.Application.Dtos.ResponseDto;
 using CafeAPI.Application.Interfaces;
 using CafeAPI.Application.Services.Abstract;
 using CafeAPI.Domain.Entities;
+using FluentValidation;
 
 namespace CafeAPI.Application.Services.Concrete
 {
     public class CategoryServices : ICategoryServices
     {
+        private readonly IValidator<CreateCategoryDto> _createCategoryValidator;
+        private readonly IValidator<UpdateCategoryDto> _updateCategoryValidator;
         private readonly IGenericRepository<Category> _categoryRepository;
         private readonly IMapper _mapper;
 
-        public CategoryServices(IGenericRepository<Category> categoryRepository, IMapper mapper)
+        public CategoryServices(IGenericRepository<Category> categoryRepository, IMapper mapper, IValidator<CreateCategoryDto> createCategoryValidator, IValidator<UpdateCategoryDto> updateCategoryValidator)
         {
             _categoryRepository = categoryRepository;
             _mapper = mapper;
+            _createCategoryValidator = createCategoryValidator;
+            _updateCategoryValidator = updateCategoryValidator;
         }
 
-        public async Task AddCategory(CreateCategoryDto dto)
+        public async Task<ResponseDto<object>> AddCategory(CreateCategoryDto dto)
         {
-            var category = _mapper.Map<Category>(dto);
-            await _categoryRepository.AddAsync(category);
+            try
+            {
+                var validate = await _createCategoryValidator.ValidateAsync(dto);
+                if (!validate.IsValid)
+                {
+                    return new ResponseDto<object>
+                    {
+                        Success = false,
+                        Data = null,
+                        Message = string.Join(" | ",validate.Errors.Select(x=>x.ErrorMessage)),
+                        ErrorCodes = ErrorCodes.ValidationError
+                    };
+                }
+                var category = _mapper.Map<Category>(dto);
+                await _categoryRepository.AddAsync(category);
+                return new ResponseDto<object> {Success = true ,Data = category, Message = "Kategori oluşturuldu."};
+            }
+            catch (Exception ex)
+            {
+                return new ResponseDto<object> {Success = false,Data = null,Message = "Bir hata oluştu!",ErrorCodes = ErrorCodes.Exception};
+
+            }
         }
 
-        public async Task DeleteCategory(int id)
+        public async Task<ResponseDto<object>> DeleteCategory(int id)
         {
-            var category = await _categoryRepository.GetByIdAsync(id);
-            await _categoryRepository.DeleteAsync(category);
+            try
+            {
+                var category = await _categoryRepository.GetByIdAsync(id);
+                if (category == null)
+                {
+                    return new ResponseDto<object> { Success = false, Data = null, Message = "Kategori bulunamadı", ErrorCodes = ErrorCodes.NotFound };
+                }
+                await _categoryRepository.DeleteAsync(category);
+                return new ResponseDto<object>{Success = true,Data = null,Message = "Kategori Silindi"};
+            }
+            catch (Exception ex)
+            {
+                return new ResponseDto<object> { Success = false, Data = null, Message = "Bir hata oluştu.", ErrorCodes = ErrorCodes.Exception };                
+            }
+            
         }
 
         public async Task<ResponseDto<List<ResultCategoryDto>>> GetAllCategories()
@@ -58,17 +96,51 @@ namespace CafeAPI.Application.Services.Concrete
            
         }
 
-        public async Task<DetailCategoryDto> GetByIdCategory(int id)
+        public async Task<ResponseDto<DetailCategoryDto>> GetByIdCategory(int id)
         {
-            var category = await _categoryRepository.GetByIdAsync(id);
-            var result = _mapper.Map<DetailCategoryDto>(category);
-            return result; 
+            try
+            {
+                var category = await _categoryRepository.GetByIdAsync(id);
+                if (category == null)
+                {
+                    return new ResponseDto<DetailCategoryDto> { Success = false,Message = "Kategori bulanamadı.",ErrorCodes = ErrorCodes.NotFound};
+                }
+                var result = _mapper.Map<DetailCategoryDto>(category);
+                return new ResponseDto<DetailCategoryDto> { Success = true, Data = result }; 
+            }
+            catch (Exception ex)
+            {
+                return new ResponseDto<DetailCategoryDto> { Success = false, Message = ex.Message, ErrorCodes = ErrorCodes.Exception };
+            }
         }
 
-        public async Task UpdateCategory(UpdateCategoryDto dto)
+        public async Task<ResponseDto<object>> UpdateCategory(UpdateCategoryDto dto)
         {
-            var category = _mapper.Map<Category>(dto);
-            await _categoryRepository.UpdateAsync(category);
+            try
+            {
+                var validate = await _updateCategoryValidator.ValidateAsync(dto);
+                if (!validate.IsValid)
+                {
+                    return new ResponseDto<object> {
+                        Success = false,
+                        Data = null,
+                        Message = string.Join(" | ", validate.Errors.Select(x => x.ErrorMessage)),
+                        ErrorCodes = ErrorCodes.ValidationError
+                    };
+                }
+                var categorydb = await _categoryRepository.GetByIdAsync(dto.Id);
+                if(categorydb == null)
+                {
+                    return new ResponseDto<object> {Success = false, Message = "Kategori bulunamadı.",ErrorCodes = ErrorCodes.NotFound};
+                }
+                var category = _mapper.Map(dto,categorydb);
+                await _categoryRepository.UpdateAsync(category);
+                return new ResponseDto<object> {Success = true,Data = null,Message = "Kategori güncellendi."};
+            }
+            catch (Exception ex) 
+            {
+                return new ResponseDto<object> {Success = false,Data = null,Message = "Bir hata oluştu", ErrorCodes = ErrorCodes.Exception};
+            }
         }
     }
 }
